+++ justfile
@@ -65,7 +65,7 @@ run:
 # run the test suite and coverage
 test:
 	python manage.py collectstatic --no-input && \
-	pytest --cov=output_explorer --cov=gateway --cov=outputs --cov=tests
+	pytest --cov=output_explorer --cov=gateway --cov=reports --cov=tests
 
 # run specific test(s)
 test-only TESTPATH:
