+++ output_explorer/settings.py
@@ -54,7 +54,7 @@ INSTALLED_APPS = [
     "django_extensions",
     "social_django",
     "gateway",
-    "outputs",
+    "reports",
 ]
 
 MIDDLEWARE = [
@@ -88,7 +88,7 @@ TEMPLATES = [
                 "social_django.context_processors.backends",
                 "social_django.context_processors.login_redirect",
                 "gateway.context_processors.show_login",
-                "outputs.context_processors.reports",
+                "reports.context_processors.reports",
             ],
         },
     },
