+++ output_explorer/urls.py
@@ -22,6 +22,6 @@ from django.urls import include, path
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("", include("social_django.urls", namespace="social")),
-    path("outputs/", include("outputs.urls", namespace="outputs")),
+    path("reports/", include("reports.urls", namespace="reports")),
     path("", include("gateway.urls", namespace="gateway")),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
