+++ reports/management/commands/ensure_superuser.py
