+++ reports/urls.py
@@ -4,7 +4,7 @@ from django.views.generic import RedirectView
 from .views import report_view
 
 
-app_name = "outputs"
+app_name = "reports"
 
 urlpatterns = [
     path("<slug:slug>/<uuid:cache_token>", report_view, name="report_view"),
