+++ tests/reports/test_admin.py
@@ -3,7 +3,7 @@ from django.contrib.auth import get_user_model
 from django.urls import reverse
 from model_bakery import baker
 
-from outputs.models import Report
+from reports.models import Report
 
 
 @pytest.mark.django_db
@@ -17,7 +17,7 @@ def test_admin_update_cache_action(client):
     report2_cache_token = report2.cache_token
 
     data = {"action": "update_cache", "_selected_action": [report1.id]}
-    change_url = reverse("admin:outputs_report_changelist")
+    change_url = reverse("admin:reports_report_changelist")
     client.login(username="admin", password="test")
     client.post(change_url, data)
